Win64 support
git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@114 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ preview.c
@@ -236,7 +236,7 @@ void recalc_preview(FilterRecordPtr pb,DIALOGREF dp){
 }
 
 OSErr drawpreview(DIALOGREF dp,void *hdc,Ptr imageptr){
-	int32 watchsusp;
+	intptr_t watchsusp;
 	OSErr e = noErr;
 	VRect srcRect;
 	UIRECT imagebounds;
